Use build flags to only build cuda or opencl versions of gominer.
-tags 'opencl' or -tags 'cuda'

Add type (cuda or opencl) to version string.

Fixes #83
