+++ README.md
@@ -2,8 +2,9 @@
 
 ## Installation
 
-You need to have the OpenCL and CUDA development libraries
-installed. You only need the runtime and drives for the one you plan
+You need to have the OpenCL or CUDA development libraries
+installed (depending on which version of gominer you would like to
+build) . You also need the runtime and drives for the one you plan
 on running (CUDA for nvidia, OpenCL for anything) To download and
 build gominer, run:
 
@@ -14,7 +15,17 @@ cd $GOPATH/src/github.com/decred
 git clone  https://github.com/decred/gominer.git
 cd gominer
 glide i
-go install $(glide nv)
+```
+
+for opencl:
+```
+go install -tags 'opencl'
+```
+
+for cuda:
+```
+make
+go install -tags 'cuda'
 ```
 
 On Ubuntu 16.04 you can install the necessary OpenCL packages (for
@@ -25,7 +36,7 @@ sudo apt-get install beignet-dev nvidia-cuda-dev nvidia-cuda-toolkit
 ```
 
 Other graphics cards will need different libraries.  We have built
-successfully on Ubuntu 16.04 with go1.6.2, g++ 5.4.0 and
+successfully on Ubuntu 16.04 with go1.6.2, go1.7.1, g++ 5.4.0 and
 beignet-dev 1.1.1-2 although other combinations should work as well.
 
 ## Running
