+++ main.go
@@ -25,7 +25,7 @@ func gominerMain() error {
 	defer backendLog.Flush()
 
 	// Show version at startup.
-	mainLog.Infof("Version %s", version())
+	mainLog.Infof("Version %s %s", version(), gpuLib())
 
 	// Enable http profiling server if requested.
 	if cfg.Profile != "" {
