+++ miner.go
@@ -45,79 +45,9 @@ func NewMiner() (*Miner, error) {
 		m.pool = s
 	}
 
-	deviceListIndex := 0
-	deviceListEnabledCount := 0
-
-	if cfg.UseCuda {
-		CUdeviceIDs, err := getCUInfo()
-		if err != nil {
-			return nil, err
-		}
-
-		// XXX Can probably combine these bits with the opencl ones once
-		// I decide what to do about the types.
-
-		for _, CUDeviceID := range CUdeviceIDs {
-			miningAllowed := false
-
-			// Enforce device restrictions if they exist
-			if len(cfg.DeviceIDs) > 0 {
-				for _, i := range cfg.DeviceIDs {
-					if deviceListIndex == i {
-						miningAllowed = true
-					}
-				}
-			} else {
-				miningAllowed = true
-			}
-
-			if miningAllowed {
-				newDevice, err := NewCuDevice(deviceListIndex, deviceListEnabledCount, CUDeviceID, m.workDone)
-				deviceListEnabledCount++
-				m.devices = append(m.devices, newDevice)
-				if err != nil {
-					return nil, err
-				}
-			}
-			deviceListIndex++
-		}
-	} else {
-		platformIDs, err := getCLPlatforms()
-		if err != nil {
-			return nil, fmt.Errorf("Could not get CL platforms: %v", err)
-		}
-
-		for p := range platformIDs {
-			platformID := platformIDs[p]
-			CLdeviceIDs, err := getCLDevices(platformID)
-			if err != nil {
-				return nil, fmt.Errorf("Could not get CL devices for platform: %v", err)
-			}
-
-			for _, CLdeviceID := range CLdeviceIDs {
-				miningAllowed := false
-
-				// Enforce device restrictions if they exist
-				if len(cfg.DeviceIDs) > 0 {
-					for _, i := range cfg.DeviceIDs {
-						if deviceListIndex == i {
-							miningAllowed = true
-						}
-					}
-				} else {
-					miningAllowed = true
-				}
-				if miningAllowed {
-					newDevice, err := NewDevice(deviceListIndex, deviceListEnabledCount, platformID, CLdeviceID, m.workDone)
-					deviceListEnabledCount++
-					m.devices = append(m.devices, newDevice)
-					if err != nil {
-						return nil, err
-					}
-				}
-				deviceListIndex++
-			}
-		}
+	m, deviceListEnabledCount, err := newMinerDevs(m)
+	if err != nil {
+		return nil, err
 	}
 
 	if deviceListEnabledCount == 0 {
